move icons to icon folder
