+++ src/components/dev-hub/icons/mdb-leaf.js
