+++ src/components/dev-hub/icons/twitter-icon.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { colorMap } from './theme';
+import { colorMap } from '../theme';
 
 const TwitterIcon = ({ color = colorMap.greyLightTwo, ...props }) => (
     <svg width="20" height="20" viewBox="0 0 512 512" {...props}>
